pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "loacl"
version = "0.1.0"
description = "Local OpenAI Assistant Chat Library"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "supabase>=2.0.0",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
    "PyJWT>=2.8.0"
]
